+++ rollup.config.js
@@ -1,6 +1,5 @@
-import babel from '@rollup/plugin-babel';
+import {babel} from '@rollup/plugin-babel';
 
-// eslint-disable-next-line import/no-anonymous-default-export
 export default [{
     input: 'src/EventTarget.js',
     output: {