+++ src/EventTarget.js
@@ -1,5 +1,56 @@
-/* eslint-disable node/no-sync, no-restricted-syntax,
-    node/no-unsupported-features/es-syntax */
+/* eslint-disable n/no-sync -- Want sync naming */
+/* eslint-disable no-restricted-syntax -- Instanceof checks */
+/* eslint-disable unicorn/no-this-assignment -- TS */
+
+/**
+ * @typedef {number} Integer
+ */
+
+/**
+ * @callback InvokeCurrentListeners
+ * @param {AllListeners} listeners
+ * @param {EventWithProps} eventCopy
+ * @param {string} type
+ * @param {boolean} [checkOnListeners]
+ * @returns {boolean}
+ */
+
+/**
+ * @typedef {{
+ *   defaultSync?: boolean,
+ *   extraProperties?: string[],
+ *   legacyOutputDidListenersThrowFlag?: boolean
+ * }} CustomOptions
+ */
+/**
+ * @typedef {{
+ *   __legacyOutputDidListenersThrowError: unknown,
+ *   target: EventTarget & {
+ *     invokeCurrentListeners: InvokeCurrentListeners,
+ *     _earlyListeners: AllListeners,
+ *     _listeners: AllListeners,
+ *     _lateListeners: AllListeners,
+ *     _defaultListeners: AllListeners
+ *   },
+ *   composed: boolean,
+ *   currentTarget: EventTarget,
+ *   eventPhase: 0|1|2|3
+ *   defaultPrevented: boolean,
+ *   type: string,
+ *   bubbles: boolean,
+ *   cancelable: boolean,
+ *   isTrusted: boolean,
+ *   timeStamp: Integer,
+ *   initEvent: (type: string, bubbles: boolean, cancelable: boolean) => void,
+ *   preventDefault: () => void,
+ *   composedPath: () => void,
+ *   detail: any,
+ *   initCustomEvent: (
+ *     type: string, canBubble: boolean, cancelable: boolean,
+ *     detail: any
+ *   ) => void
+ * }} EventWithProps
+ */
 
 // Todo: Switch to ES6 classes
 
@@ -11,9 +62,17 @@ const phases = {
 };
 
 const ShimDOMException = typeof DOMException === 'undefined'
-    // Todo: Better polyfill (if even needed here)
-    // eslint-disable-next-line no-shadow
-    ? function DOMException (msg, name) { // No need for `toString` as same as for `Error`
+// Todo: Better polyfill (if even needed here)
+/* eslint-disable no-shadow -- Polyfill */
+    // eslint-disable-next-line operator-linebreak -- TS/JSDoc needs
+    ?
+    /**
+     * @param {string} msg
+     * @param {string} name
+     * @returns {Error}
+     */
+    function DOMException (msg, name) { // No need for `toString` as same as for `Error`
+        /* eslint-enable no-shadow -- Polyfill */
         const err = new Error(msg);
         err.name = name;
         return err;
@@ -25,7 +84,8 @@ const evCfg = new WeakMap();
 
 // Todo: Set _ev argument outside of this function
 
-/* eslint-disable func-name-matching, no-shadow */
+/* eslint-disable func-name-matching -- Shim vs. Polyfill */
+/* eslint-disable no-shadow -- Polyfilling */
 /**
 * We use an adapter class rather than a proxy not only for compatibility
 * but also since we have to clone native event properties anyways in order
@@ -35,14 +95,16 @@ const evCfg = new WeakMap();
 * @class
 * @param {string} type
 */
-const ShimEvent = function Event (type) {
-    /* eslint-enable func-name-matching, no-shadow */
+const ShimEvent = /** @type {unknown} */ function Event (type) {
+    /* eslint-enable func-name-matching -- Shim vs. Polyfill */
+    /* eslint-enable no-shadow -- Polyfilling */
     // For WebIDL checks of function's `length`, we check `arguments` for the optional arguments
+    // @ts-expect-error
     this[Symbol.toStringTag] = 'Event';
     this.toString = () => {
         return '[object Event]';
     };
-    // eslint-disable-next-line prefer-rest-params
+    // eslint-disable-next-line prefer-rest-params -- Don't want to change signature
     let [, evInit, _ev] = arguments;
     if (!arguments.length) {
         throw new TypeError("Failed to construct 'Event': 1 argument required, but only 0 present.");
@@ -50,6 +112,7 @@ const ShimEvent = function Event (type) {
     evInit = evInit || {};
     _ev = _ev || {};
 
+    /** @type {EventWithProps} */
     const _evCfg = {};
     if ('composed' in evInit) {
         _evCfg.composed = evInit.composed;
@@ -60,29 +123,31 @@ const ShimEvent = function Event (type) {
 
     ev.set(this, _ev);
     evCfg.set(this, _evCfg);
-    this.initEvent(type, evInit.bubbles, evInit.cancelable);
-    Object.defineProperties(
-        this,
-        ['target', 'currentTarget', 'eventPhase', 'defaultPrevented'].reduce((obj, prop) => {
-            obj[prop] = {
-                get () {
-                    return (/* prop in _evCfg && */ _evCfg[prop] !== undefined)
-                        ? _evCfg[prop]
-                        : (
-                            prop in _ev
-                                ? _ev[prop]
-                                : (
-                                    // Defaults
-                                    prop === 'eventPhase'
-                                        ? 0
-                                        : (prop === 'defaultPrevented' ? false : null)
-                                )
-                        );
-                }
-            };
-            return obj;
-        }, {})
-    );
+    const that = /** @type {unknown} */ (this);
+    /** @type {ShimEvent} */ (that).initEvent(type, evInit.bubbles, evInit.cancelable);
+
+    ['target', 'currentTarget', 'eventPhase', 'defaultPrevented'].forEach((pr) => {
+        const prop = /** @type {"target"|"currentTarget"|"eventPhase"|"defaultPrevented"} */ (
+            pr
+        );
+        Object.defineProperty(this, prop, {
+            get () {
+                return (/* prop in _evCfg && */ _evCfg[prop] !== undefined)
+                    ? _evCfg[prop]
+                    : (
+                        prop in _ev
+                            ? _ev[prop]
+                            : (
+                                // Defaults
+                                prop === 'eventPhase'
+                                    ? 0
+                                    : (prop === 'defaultPrevented' ? false : null)
+                            )
+                    );
+            }
+        });
+    });
+
     const props = [
         // Event
         'type',
@@ -96,7 +161,16 @@ const ShimEvent = function Event (type) {
         props.push('detail', 'initCustomEvent');
     }
 
-    Object.defineProperties(this, props.reduce((obj, prop) => {
+    Object.defineProperties(this, props.reduce((obj, pr) => {
+        const prop =
+            /**
+             * @type {"type"|"bubbles"|"cancelable"|"isTrusted"|
+             *   "timeStamp"|"initEvent"|"composedPath"|"composed"|
+             *   "detail"|"initCustomEvent"
+             * }
+             */ (
+                pr
+            );
         obj[prop] = {
             get () {
                 return prop in _evCfg
@@ -111,10 +185,12 @@ const ShimEvent = function Event (type) {
             }
         };
         return obj;
-    }, {}));
+    }, /** @type {{[key: string]: any}} */ ({})));
 };
 
+// @ts-expect-error Casting doesn't work
 ShimEvent.prototype.preventDefault = function () {
+    // @ts-expect-error Needed for exporting
     if (!(this instanceof ShimEvent)) {
         throw new TypeError('Illegal invocation');
     }
@@ -127,14 +203,20 @@ ShimEvent.prototype.preventDefault = function () {
         }
     }
 };
+
+// @ts-expect-error Casting doesn't work
 ShimEvent.prototype.stopImmediatePropagation = function () {
     const _evCfg = evCfg.get(this);
     _evCfg._stopImmediatePropagation = true;
 };
+
+// @ts-expect-error Casting doesn't work
 ShimEvent.prototype.stopPropagation = function () {
     const _evCfg = evCfg.get(this);
     _evCfg._stopPropagation = true;
 };
+
+// @ts-expect-error Casting doesn't work
 ShimEvent.prototype.initEvent = function (type, bubbles, cancelable) { // Chrome currently has function length 1 only but WebIDL says 3
     // const bubbles = arguments[1];
     // const cancelable = arguments[2];
@@ -143,6 +225,30 @@ ShimEvent.prototype.initEvent = function (type, bubbles, cancelable) { // Chrome
     if (_evCfg._dispatched) {
         return;
     }
+
+    Object.defineProperty(this, 'type', {
+        enumerable: true,
+        configurable: true,
+        get () {
+            return type;
+        }
+    });
+
+    Object.defineProperty(this, 'bubbles', {
+        enumerable: true,
+        configurable: true,
+        get () {
+            return bubbles;
+        }
+    });
+
+    Object.defineProperty(this, 'cancelable', {
+        enumerable: true,
+        configurable: true,
+        get () {
+            return cancelable;
+        }
+    });
 
     _evCfg.type = type;
     if (bubbles !== undefined) {
@@ -153,6 +259,7 @@ ShimEvent.prototype.initEvent = function (type, bubbles, cancelable) { // Chrome
     }
 };
 ['type', 'target', 'currentTarget'].forEach((prop) => {
+    // @ts-expect-error Casting doesn't work
     Object.defineProperty(ShimEvent.prototype, prop, {
         enumerable: true,
         configurable: true,
@@ -162,6 +269,7 @@ ShimEvent.prototype.initEvent = function (type, bubbles, cancelable) { // Chrome
     });
 });
 ['eventPhase', 'defaultPrevented', 'bubbles', 'cancelable', 'timeStamp'].forEach((prop) => {
+    // @ts-expect-error Casting doesn't work
     Object.defineProperty(ShimEvent.prototype, prop, {
         enumerable: true,
         configurable: true,
@@ -176,34 +284,43 @@ ShimEvent.prototype.initEvent = function (type, bubbles, cancelable) { // Chrome
         writable: false,
         value: i
     });
+    // @ts-expect-error Casting doesn't work
     Object.defineProperty(ShimEvent.prototype, prop, {
         writable: false,
         value: i
     });
 });
+// @ts-expect-error Casting doesn't work
 ShimEvent[Symbol.toStringTag] = 'Function';
+
+// @ts-expect-error Casting doesn't work
 ShimEvent.prototype[Symbol.toStringTag] = 'EventPrototype';
 Object.defineProperty(ShimEvent, 'prototype', {
     writable: false
 });
 
-/* eslint-disable func-name-matching, no-shadow */
+/* eslint-disable func-name-matching -- Polyfill */
+/* eslint-disable no-shadow -- Polyfill */
 /**
- *
- * @param {string} type
  * @class
+ * @param {string} type
  */
-const ShimCustomEvent = function CustomEvent (type) {
-    /* eslint-enable func-name-matching, no-shadow */
-    // eslint-disable-next-line prefer-rest-params
+const ShimCustomEvent = /** @type {unknown} */ function CustomEvent (type) {
+    /* eslint-enable func-name-matching -- Polyfill */
+    /* eslint-enable no-shadow -- Polyfill */
+
+    // eslint-disable-next-line prefer-rest-params -- Keep signature
     let [, evInit, _ev] = arguments;
+    // @ts-expect-error Casting doesn't work
     ShimEvent.call(this, type, evInit, _ev);
+    // @ts-expect-error
     this[Symbol.toStringTag] = 'CustomEvent';
     this.toString = () => {
         return '[object CustomEvent]';
     };
     // var _evCfg = evCfg.get(this);
     evInit = evInit || {};
+    // @ts-ignore
     this.initCustomEvent(
         type,
         evInit.bubbles,
@@ -211,20 +328,24 @@ const ShimCustomEvent = function CustomEvent (type) {
         'detail' in evInit ? evInit.detail : null
     );
 };
-Object.defineProperty(ShimCustomEvent.prototype, 'constructor', {
+// @ts-expect-error Casting doesn't work
+Object.defineProperty((ShimCustomEvent).prototype, 'constructor', {
     enumerable: false,
     writable: true,
     configurable: true,
     value: ShimCustomEvent
 });
+// @ts-expect-error Casting doesn't work
 ShimCustomEvent.prototype.initCustomEvent = function (type, bubbles, cancelable, detail) {
+    // @ts-expect-error Needed for exporting
     if (!(this instanceof ShimCustomEvent)) {
         throw new TypeError('Illegal invocation');
     }
     const _evCfg = evCfg.get(this);
+    // @ts-expect-error Casting doesn't work
     ShimCustomEvent.call(this, type, {
         bubbles, cancelable, detail
-    // eslint-disable-next-line prefer-rest-params
+    // eslint-disable-next-line prefer-rest-params -- Keep signature
     }, arguments[4]);
 
     if (_evCfg._dispatched) {
@@ -240,9 +361,12 @@ ShimCustomEvent.prototype.initCustomEvent = function (type, bubbles, cancelable,
         }
     });
 };
+// @ts-expect-error Casting doesn't work
 ShimCustomEvent[Symbol.toStringTag] = 'Function';
+// @ts-expect-error Casting doesn't work
 ShimCustomEvent.prototype[Symbol.toStringTag] = 'CustomEventPrototype';
 
+// @ts-expect-error Casting doesn't work
 Object.defineProperty(ShimCustomEvent.prototype, 'detail', {
     enumerable: true,
     configurable: true,
@@ -256,34 +380,36 @@ Object.defineProperty(ShimCustomEvent, 'prototype', {
 
 /**
  *
- * @param {Event} e
- * @returns {ShimEvent}
+ * @param {EventWithProps} e
+ * @returns {EventWithProps}
  */
 function copyEvent (e) {
     const {bubbles, cancelable, detail, type} = e;
     if ('detail' in e) {
+        // @ts-expect-error Casting doesn't work
         return new ShimCustomEvent(
             type, {bubbles, cancelable, detail}, e
         );
     }
+    // @ts-expect-error Casting doesn't work
     return new ShimEvent(type, {bubbles, cancelable}, e);
 }
 
 /**
-* @typedef {PlainObject} ListenerOptions
-* @property {boolean} once Remove listener after invoking once
-* @property {boolean} passive Don't allow `preventDefault`
-* @property {boolean} capture Use `_children` and set `eventPhase`
+* @typedef {object} ListenerOptions
+* @property {boolean} [once] Remove listener after invoking once
+* @property {boolean} [passive] Don't allow `preventDefault`
+* @property {boolean} [capture] Use `_children` and set `eventPhase`
 */
 
 /**
-* @typedef {PlainObject} ListenerAndOptions
-* @property {listener} listener
+* @typedef {object} ListenerAndOptions
+* @property {Listener} listener
 * @property {ListenerOptions} options
 */
 
 /**
-* @typedef {PlainObject} ListenerInfo
+* @typedef {object} ListenerInfo
 * @property {ListenerAndOptions[]} listenersByTypeOptions
 * @property {ListenerOptions} options
 * @property {ListenerAndOptions[]} listenersByType
@@ -291,18 +417,24 @@ function copyEvent (e) {
 
 /**
 * @callback Listener
-* @param {Event} e
+* @param {EventWithProps} e
 * @returns {boolean}
 */
 
 /**
  * Keys are event types.
- * @typedef {Object<string,Listener[]>} Listeners
-*/
+ * @typedef {{[key: string]: Listener[]}} Listeners
+ */
+
+/**
+ * @typedef {{
+ *   [type: string]: ListenerAndOptions[]
+ * }} AllListeners
+ */
 
 /**
  *
- * @param {Listeners} listeners
+ * @param {AllListeners} listeners
  * @param {string} type
  * @param {boolean|ListenerOptions} options
  * @returns {ListenerInfo}
@@ -310,20 +442,27 @@ function copyEvent (e) {
 function getListenersOptions (listeners, type, options) {
     let listenersByType = listeners[type];
     if (listenersByType === undefined) listeners[type] = listenersByType = [];
-    options = typeof options === 'boolean' ? {capture: options} : (options || {});
-    const stringifiedOptions = JSON.stringify(options);
+    const opts = typeof options === 'boolean' ? {capture: options} : (options || {});
+    const stringifiedOptions = JSON.stringify(opts);
     const listenersByTypeOptions = listenersByType.filter((obj) => {
         return stringifiedOptions === JSON.stringify(obj.options);
     });
-    return {listenersByTypeOptions, options, listenersByType};
+    return {listenersByTypeOptions, options: opts, listenersByType};
 }
 
 const methods = {
+    /**
+     * @param {AllListeners} listeners
+     * @param {Listener} listener
+     * @param {string} type
+     * @param {boolean|ListenerOptions} options
+     * @returns {void}
+     */
     addListener (listeners, listener, type, options) {
-        const listenerOptions = getListenersOptions(listeners, type, options);
-        const {listenersByTypeOptions} = listenerOptions;
-        ({options} = listenerOptions);
-        const {listenersByType} = listenerOptions;
+        const listenersOptions = getListenersOptions(listeners, type, options);
+        const {listenersByTypeOptions} = listenersOptions;
+        ({options} = listenersOptions);
+        const {listenersByType} = listenersOptions;
 
         if (listenersByTypeOptions.some((l) => {
             return l.listener === listener;
@@ -331,10 +470,17 @@ const methods = {
         listenersByType.push({listener, options});
     },
 
+    /**
+     * @param {AllListeners} listeners
+     * @param {Listener} listener
+     * @param {string} type
+     * @param {boolean|ListenerOptions} options
+     * @returns {void}
+     */
     removeListener (listeners, listener, type, options) {
-        const listenerOptions = getListenersOptions(listeners, type, options);
-        const {listenersByType} = listenerOptions;
-        const stringifiedOptions = JSON.stringify(listenerOptions.options);
+        const listenersOptions = getListenersOptions(listeners, type, options);
+        const {listenersByType} = listenersOptions;
+        const stringifiedOptions = JSON.stringify(listenersOptions.options);
 
         listenersByType.some((l, i) => {
             if (l.listener === listener && stringifiedOptions === JSON.stringify(l.options)) {
@@ -346,56 +492,101 @@ const methods = {
         });
     },
 
+    /**
+     *
+     * @param {AllListeners} listeners
+     * @param {Listener} listener
+     * @param {string} type
+     * @param {boolean|ListenerOptions} options
+     * @returns {boolean}
+     */
     hasListener (listeners, listener, type, options) {
-        const listenerOptions = getListenersOptions(listeners, type, options);
-        const {listenersByTypeOptions} = listenerOptions;
+        const listenersOptions = getListenersOptions(listeners, type, options);
+        const {listenersByTypeOptions} = listenersOptions;
         return listenersByTypeOptions.some((l) => {
             return l.listener === listener;
         });
     }
 };
 
-/* eslint-disable no-shadow */
+/* eslint-disable no-shadow -- Polyfill */
 /**
  * @class
  */
 function EventTarget () {
-    /* eslint-enable no-shadow */
+    /* eslint-enable no-shadow -- Polyfill */
     throw new TypeError('Illegal constructor');
 }
 
-Object.assign(EventTarget.prototype, ['Early', '', 'Late', 'Default'].reduce(function (obj, listenerType) {
+/**
+ * @typedef {"addEarlyEventListener"|"removeEarlyEventListener"|"hasEarlyEventListener"|
+ *   "addEventListener"|"removeEventListener"|"hasEventListener"|
+ *   "addLateEventListener"|"removeLateEventListener"|"hasLateEventListener"|
+ *   "addDefaultEventListener"|"removeDefaultEventListener"|"hasDefaultEventListener"
+ * } ListenerName
+ */
+Object.assign(EventTarget.prototype, ['Early', '', 'Late', 'Default'].reduce(function (
+    /** @type {{[key: string]: Function}} */
+    obj,
+    listenerType
+) {
     ['add', 'remove', 'has'].forEach(function (method) {
-        obj[method + listenerType + 'EventListener'] = function (type, listener) {
-            // eslint-disable-next-line prefer-rest-params
+        const mainMethod = /** @type {ListenerName} */ (method + listenerType + 'EventListener');
+        /**
+         * @param {string} type
+         * @param {Listener|{handleEvent: Listener}} listener
+         * @this {EventTarget & {
+         *   _earlyListeners: AllListeners,
+         *   _listeners: AllListeners,
+         *   _lateListeners: AllListeners,
+         *   _defaultListeners: AllListeners,
+         * }}
+         * @returns {boolean|void}
+         */
+        obj[mainMethod] = function (type, listener) {
+            // eslint-disable-next-line prefer-rest-params -- Keep signature
             const options = arguments[2]; // We keep the listener `length` as per WebIDL
             if (arguments.length < 2) throw new TypeError('2 or more arguments required');
             if (typeof type !== 'string') {
+                // @ts-expect-error It's ok to construct
                 throw new ShimDOMException('UNSPECIFIED_EVENT_TYPE_ERR', 'UNSPECIFIED_EVENT_TYPE_ERR');
             }
             try {
                 // As per code such as the following, handleEvent may throw,
                 //  but is uncaught
                 // https://github.com/web-platform-tests/wpt/blob/master/IndexedDB/fire-error-event-exception.html#L54-L56
-                if (listener.handleEvent && listener.handleEvent.bind) {
+                if ('handleEvent' in listener && listener.handleEvent.bind) {
                     listener = listener.handleEvent.bind(listener);
                 }
             } catch (err) {
-                // eslint-disable-next-line no-console
+                // eslint-disable-next-line no-console -- Feedback to user
                 console.log('Uncaught `handleEvent` error', err);
             }
 
-            const arrStr = '_' + listenerType.toLowerCase() + (listenerType === '' ? 'l' : 'L') + 'isteners';
+            const arrStr = /** @type {"_earlyListeners"|"_listeners"|"_lateListeners"|"_defaultListeners"} */ (
+                '_' + listenerType.toLowerCase() + (listenerType === '' ? 'l' : 'L') + 'isteners'
+            );
             if (!this[arrStr]) {
                 Object.defineProperty(this, arrStr, {value: {}});
             }
-            return methods[method + 'Listener'](this[arrStr], listener, type, options);
+            const meth = /** @type {"addListener"|"removeListener"|"hasListener"} */ (
+                method + 'Listener'
+            );
+            return methods[meth](
+                this[arrStr], /** @type {Listener} */ (listener), type, options
+            );
         };
     });
     return obj;
 }, {}));
 
 Object.assign(EventTarget.prototype, {
+    _legacyOutputDidListenersThrowCheck: undefined,
+    /**
+     * @param {CustomOptions} customOptions
+     * @this {EventTarget.prototype}
+     * @returns {void}
+     */
     __setOptions (customOptions) {
         customOptions = customOptions || {};
         // Todo: Make into event properties?
@@ -406,12 +597,32 @@ Object.assign(EventTarget.prototype, {
             this._extraProperties.push('__legacyOutputDidListenersThrowError');
         }
     },
+    /**
+     * @param {ShimEvent} e
+     * @this {EventTarget & {
+     *   _dispatchEvent: (e: ShimEvent|ShimCustomEvent, setTarget: boolean) => boolean,
+    * }}
+     * @returns {boolean}
+     */
     dispatchEvent (e) {
         return this._dispatchEvent(e, true);
     },
+    /**
+     * @param {EventWithProps} e
+     * @param {boolean} setTarget
+     * @this {EventTarget.prototype & {
+     *   _earlyListeners: AllListeners,
+     *   _listeners: AllListeners,
+     *   _lateListeners: AllListeners,
+     *   _defaultListeners: AllListeners,
+     * }}
+     * @returns {boolean}
+     */
     _dispatchEvent (e, setTarget) {
         ['early', '', 'late', 'default'].forEach((listenerType) => {
-            const arrStr = '_' + listenerType + (listenerType === '' ? 'l' : 'L') + 'isteners';
+            const arrStr = /** @type {"_earlyListeners"|"_listeners"|"_lateListeners"|"_defaultListeners"} */ (
+                '_' + listenerType + (listenerType === '' ? 'l' : 'L') + 'isteners'
+            );
             if (!this[arrStr]) {
                 Object.defineProperty(this, arrStr, {value: {}});
             }
@@ -419,9 +630,11 @@ Object.assign(EventTarget.prototype, {
 
         let _evCfg = evCfg.get(e);
         if (_evCfg && setTarget && _evCfg._dispatched) {
+            // @ts-expect-error It's ok to construct
             throw new ShimDOMException('The object is in an invalid state.', 'InvalidStateError');
         }
 
+        /** @type {EventWithProps} */
         let eventCopy;
         if (_evCfg) {
             eventCopy = e;
@@ -429,9 +642,12 @@ Object.assign(EventTarget.prototype, {
             eventCopy = copyEvent(e);
             _evCfg = evCfg.get(eventCopy);
             _evCfg._dispatched = true;
-            this._extraProperties.forEach((prop) => {
+
+            /** @type {string[]} */
+            (this._extraProperties).forEach((prop) => {
                 if (prop in e) {
-                    eventCopy[prop] = e[prop]; // Todo: Put internal to `ShimEvent`?
+                    /** @type {{[key: string]: any}} */ (eventCopy)[prop] =
+                        /** @type {{[key: string]: any}} */ (e)[prop]; // Todo: Put internal to `ShimEvent`?
                 }
             });
         }
@@ -484,7 +700,7 @@ Object.assign(EventTarget.prototype, {
 
         if (setTarget) _evCfg.target = this;
 
-        switch (eventCopy.eventPhase) {
+        switch ('eventPhase' in eventCopy && eventCopy.eventPhase) {
         case phases.CAPTURING_PHASE: {
             if (_evCfg._stopPropagation) {
                 return continueEventDispatch();
@@ -522,15 +738,15 @@ Object.assign(EventTarget.prototype, {
         default: {
             _evCfg.eventPhase = phases.AT_TARGET; // Temporarily set before we invoke early listeners
             this.invokeCurrentListeners(this._earlyListeners, eventCopy, type);
-            if (!this.__getParent) {
+            if (!('__getParent' in this)) {
                 _evCfg.eventPhase = phases.AT_TARGET;
                 return this._dispatchEvent(eventCopy, false);
             }
 
-            /* eslint-disable consistent-this */
+            /* eslint-disable consistent-this -- Readability */
             let par = this;
             let root_ = this;
-            /* eslint-enable consistent-this */
+            /* eslint-enable consistent-this -- Readability */
             while (par.__getParent && (par = par.__getParent()) !== null) {
                 if (!_evCfg._children) {
                     _evCfg._children = [];
@@ -544,11 +760,17 @@ Object.assign(EventTarget.prototype, {
         }
         }
     },
+
+    /**
+     * @type {InvokeCurrentListeners}
+     * @this {EventTarget.prototype & {[key: string]: Listener}}
+     */
     invokeCurrentListeners (listeners, eventCopy, type, checkOnListeners) {
         const _evCfg = evCfg.get(eventCopy);
         _evCfg.currentTarget = this;
 
         const listOpts = getListenersOptions(listeners, type, {});
+        // eslint-disable-next-line unicorn/prefer-spread -- Performance?
         const listenersByType = listOpts.listenersByType.concat();
         const dummyIPos = listenersByType.length ? 1 : 0;
 
@@ -600,19 +822,31 @@ Object.assign(EventTarget.prototype, {
 
         return !eventCopy.defaultPrevented;
     },
-    // eslint-disable-next-line promise/prefer-await-to-callbacks
+    /* eslint-disable promise/prefer-await-to-callbacks -- Try-catch */
+    /**
+     * @param {EventWithProps} evt
+     * @param {() => void} cb
+     * @returns {void}
+     */
     tryCatch (evt, cb) {
+        /* eslint-enable promise/prefer-await-to-callbacks -- Try-catch */
         try {
             // Per MDN: Exceptions thrown by event handlers are reported
             //    as uncaught exceptions; the event handlers run on a nested
             //    callstack: they block the caller until they complete, but
             //    exceptions do not propagate to the caller.
-            // eslint-disable-next-line promise/prefer-await-to-callbacks, node/callback-return
+            // eslint-disable-next-line promise/prefer-await-to-callbacks, n/callback-return --  Try-catch
             cb();
         } catch (err) {
             this.triggerErrorEvent(err, evt);
         }
     },
+
+    /**
+     * @param {unknown} err
+     * @param {EventWithProps} evt
+     * @returns {void}
+     */
     triggerErrorEvent (err, evt) {
         let error = err;
         if (typeof err === 'string') {
@@ -645,11 +879,17 @@ Object.assign(EventTarget.prototype, {
                 // `ErrorEvent` properly triggers `window.onerror` and `window.addEventListener('error')` handlers
                 const errEv = new ErrorEvent('error', {
                     error: err,
-                    message: error.message || '',
+                    message: /** @type {Error} */ (error).message || '',
                     // We can't get the actually useful user's values!
-                    filename: error.fileName || '',
-                    lineno: error.lineNumber || 0,
-                    colno: error.columnNumber || 0
+                    filename: /** @type {Error & {fileName: string}} */ (
+                        error
+                    ).fileName || '',
+                    lineno: /** @type {Error & {lineNumber: Integer}} */ (
+                        error
+                    ).lineNumber || 0,
+                    colno: /** @type {Error & {columnNumber: Integer}} */ (
+                        error
+                    ).columnNumber || 0
                 });
                 window.dispatchEvent(errEv);
                 // console.log(err); // Should we auto-log for user?
@@ -679,17 +919,27 @@ Object.defineProperty(EventTarget, 'prototype', {
 
 const ShimEventTarget = EventTarget;
 const EventTargetFactory = {
+
+    /**
+     * @param {CustomOptions} customOptions
+     * @returns {EventTarget}
+     */
     createInstance (customOptions) {
-        /* eslint-disable no-shadow */
+        /* eslint-disable func-name-matching -- Shim vs. Polyfill */
+        /* eslint-disable no-shadow -- Polyfill */
         /**
          * @class
+         * @this {typeof ShimEventTarget.prototype}
          */
-        function EventTarget () {
-            /* eslint-enable no-shadow */
+        const ET = /** @type {unknown} */ function EventTarget () {
+            /* eslint-enable no-shadow -- Polyfill */
+            /* eslint-enable func-name-matching -- Shim vs. Polyfill */
             this.__setOptions(customOptions);
-        }
-        EventTarget.prototype = ShimEventTarget.prototype;
-        return new EventTarget();
+        };
+        // @ts-expect-error Casting doesn't work
+        ET.prototype = ShimEventTarget.prototype;
+        // @ts-expect-error Casting doesn't work
+        return new ET();
     }
 };
 
@@ -704,7 +954,8 @@ EventTarget.EventTargetFactory = EventTargetFactory;
  */
 function setPrototypeOfCustomEvent () {
     // TODO: IDL needs but reported as slow!
-    Object.setPrototypeOf(ShimCustomEvent, ShimEvent);
+    Object.setPrototypeOf(ShimCustomEvent, /** @type {object} */ (ShimEvent));
+    // @ts-expect-error How to overcome?
     Object.setPrototypeOf(ShimCustomEvent.prototype, ShimEvent.prototype);
 }